src/requirements.txt
celery[redis]
influxdb-client
paho-mqtt
